+++ Source/OxyPlot/Series/AreaSeries.cs
@@ -240,7 +240,7 @@ namespace OxyPlot.Series
             double minDistSquared = this.MinimumSegmentLength * this.MinimumSegmentLength;
 
             var clippingRect = this.GetClippingRect();
-            rc.SetClip(clippingRect);
+            using var _ = rc.AutoResetClip(clippingRect);
 
             var areaContext = new AreaRenderContext
             {
@@ -266,7 +266,6 @@ namespace OxyPlot.Series
 
             if (chunksOfPoints.Count != chunksOfPoints2.Count)
             {
-                rc.ResetClip();
                 return;
             }
 
@@ -317,8 +316,6 @@ namespace OxyPlot.Series
                     this.EdgeRenderingMode,
                     1);
             }
-
-            rc.ResetClip();
         }
 
         /// <summary>