+++ Source/OxyPlot/Svg/SvgRenderContext.cs
@@ -29,11 +29,6 @@ namespace OxyPlot
         private bool disposed;
 
         /// <summary>
-        /// The current clipping rectangle.
-        /// </summary>
-        private OxyRect? clipRect;
-
-        /// <summary>
         /// Initializes a new instance of the <see cref="SvgRenderContext" /> class.
         /// </summary>
         /// <param name="s">The stream.</param>
@@ -267,38 +262,6 @@ namespace OxyPlot
             this.w.WriteImage(srcX, srcY, srcWidth, srcHeight, destX, destY, destWidth, destHeight, source);
         }
 
-        /// <inheritdoc/>
-        public override bool SetClip(OxyRect rect)
-        {
-            if (this.clipRect != null)
-            {
-                if (rect.Equals(this.clipRect.Value))
-                {
-                    return true;
-                }
-                else
-                {
-                    this.ResetClip();
-                }
-            }
-
-            this.clipRect = rect;
-            this.w.BeginClip(rect.Left, rect.Top, rect.Width, rect.Height);
-            return true;
-        }
-
-        /// <inheritdoc/>
-        public override void ResetClip()
-        {
-            if (this.clipRect == null)
-            {
-                return;
-            }
-
-            this.clipRect = null;
-            this.w.EndClip();
-        }
-
         /// <summary>
         /// Releases unmanaged and - optionally - managed resources
         /// </summary>
@@ -314,6 +277,18 @@ namespace OxyPlot
             }
 
             this.disposed = true;
+        }
+
+        /// <inheritdoc/>
+        protected override void SetClip(OxyRect clippingRectangle)
+        {
+            this.w.BeginClip(clippingRectangle.Left, clippingRectangle.Top, clippingRectangle.Width, clippingRectangle.Height);
+        }
+
+        /// <inheritdoc/>
+        protected override void ResetClip()
+        {
+            this.w.EndClip();
         }
     }
 }